Added MinHealth, restarts scene, and fixed patterns
I added some code to make the game restart when the health reaches 0. I also realized I was had one pattern appearing 2/3 of the time, when each pattern should occur approximately 1/3 of the time if it is truly random. This made me realize that ObstaclePattern3 was an exact copy of ObstaclePattern2 so I altered a Y coordinate of OP3 to fill in the missing pattern.